eda and hub entry points
