+++ frontend/awx/Awx.tsx
@@ -1,6 +1,6 @@
 import { lazy } from 'react';
 import { createRoot } from 'react-dom/client';
-const AwxMain = lazy(() => import('./AwxMain'));
+const Main = lazy(() => import('./AwxMain'));
 document.body.innerHTML = '<div id="app"></div>';
 const root = createRoot(document.getElementById('app')!);
-root.render(<AwxMain />);
+root.render(<Main />);
